requirements.txt
# Core pipeline
moviepy==1.0.3
Pillow==10.4.0
numpy==2.0.2
pyyaml==6.0.2
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.5

# Script generation
ollama==0.3.3

# Image generation
replicate==0.31.0          # Hybrid mode: Replicate API
# comfyui is run as a separate local server, no pip install needed

# Voice / TTS
kokoro-onnx==0.3.5
soundfile==0.12.1

# Music generation
# audiocraft==1.3.0 pins torch==2.1.0 which has no Python 3.12 wheel.
# Install separately: pip install audiocraft==1.3.0 --no-deps
torch>=2.2.0

# Upload
google-api-python-client==2.139.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Optional: Claude API fallback for script generation
anthropic==0.34.2
